# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Dental Backend API
  version: 1.0.0
  description: OpenAPI specification for dental clinic backend (services, appointments, payments, users, admin analytics, auth, chat)
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/services:
    get:
      summary: List all services
      security: []
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ServiceResponseList'
    post:
      summary: Create a service (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '200':
          description: Created service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ServiceResponse'
  /api/services/{id}:
    get:
      summary: Get service by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ServiceResponse'
    put:
      summary: Update service (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: Updated service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ServiceResponse'
    delete:
      summary: Delete service (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
...
  /api/payments/record:
    post:
      summary: Record payment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: query
          required: true
          schema:
            type: integer
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: method
          in: query
          required: true
          schema:
            type: string
        - name: invoiceNumber
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Payment'
  /api/payments/appointment/{appointmentId}:
    get:
      summary: Get payments for appointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Payments list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaymentList'
  /api/payments/{paymentId}:
    get:
      summary: Get payment by id
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Payment'
  /api/users/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
  /api/users/me/preferences:
    patch:
      summary: Update current user preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
  /api/users:
    post:
      summary: Create user (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEntity'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
  /api/users/{id}:
    put:
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEntity'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
    delete:
      summary: Delete user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/users/{id}/debt:
    patch:
      summary: Update user debt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: debt
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Debt updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
  /api/users/{id}/service-status:
    patch:
      summary: Update user service status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: serviceStatus
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
...
  /api/users/{id}/role:
    patch:
      summary: Change user's role (admin only)
      description: Invalidate user's refresh tokens after role change so the user must re-authenticate.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: role
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/UserRole'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserEntity'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ApiResponse_String:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: string
...
    ServiceResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        durationMinutes:
          type: integer
    ApiResponse_ServiceResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ServiceResponse'
    ApiResponse_ServiceResponseList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ServiceResponse'
    CreateServiceRequest:
      type: object
      required: [name, price, durationMinutes]
      properties:
        name:
          type: string
        price:
          type: number
        description:
          type: string
        durationMinutes:
          type: integer
    UpdateServiceRequest:
      $ref: '#/components/schemas/CreateServiceRequest'
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        amount:
          type: number
          format: double
        method:
          type: string
        invoiceNumber:
          type: string
        paidAt:
          type: string
          format: date-time
    ApiResponse_Payment:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Payment'
    ApiResponse_PaymentList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Payment'
    UserEntity:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        fullName:
          type: string
        emailVerified:
          type: boolean
        enabled:
          type: boolean
        themePreference:
          type: string
        languagePreference:
          type: string
        notificationPreference:
          type: string
        debt:
          type: number
          format: double
        serviceStatus:
          type: string
    ApiResponse_UserEntity:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserEntity'
    UpdatePreferencesRequest:
      type: object
      properties:
        themePreference:
          type: string
        languagePreference:
          type: string
        notificationPreference:
          type: string
...
    UserRole:
      type: string
      enum:
        - CUSTOMER
        - RECEPTIONIST
        - DENTIST
        - ADMIN
        - PATIENT